Little fixes: Fixed buffer dimensions not being recalculated even if the exceeded screen space. Removed unnecesary dependencies. Now pipeline.c is obsolete, but I'll leave it anyways. Next up is moving swap chain into pipeline, zBuffer fast checks

+++ ConsoleApplication5/GouraudPS.c
@@ -8,7 +8,7 @@ extern cBuf_GouraudVS GVSCB;
 
 CHAR_INFO virtual(Apply)(void* self, void* _in)
 {
-	return Pick(BG_Ditry_Gray, _mm_cvtss_f32(((VMVECTOR*)_in)[1]) * 4.0f, GVSCB.diffuse);
+	return Pick(BG_Ditry_Gray, (Byte)(_mm_cvtss_f32(((VMVECTOR*)_in)[1]) * 4.0f), GVSCB.diffuse);
 }
 
 Constructor(selfptr, va_list *ap)
